@@ -2,24 +2,29 @@ class WebUser < ActiveRecord::Base
   attr_accessor :password
   before_save :set_hashed_password
 
-  validates :email, :format => { :with => /@/ }
-  validates :email, :uniqueness => { :message => 'is already in use.' }
-  validates :password, :confirmation => true
+  #TODO: better email validation?
+  validates :email, format: { :with => /@/ }
+  validates :email, uniqueness: { message: 'is already in use.' }
+  validates :password, confirmation: true
 
   def set_hashed_password
-    return if password.blank?
-    self.password_salt = SecureRandom.urlsafe_base64
-    self.password_hash = self.hash_password(self.password)
+    if password.blank?
+      self.password_salt = nil
+      self.password_hash = nil
+    else
+      self.password_salt = SecureRandom.urlsafe_base64
+      self.password_hash = self.hash_password(self.password)
+    end
   end
 
   def hash_password(password)
     Digest::SHA512.hexdigest(password + self.password_salt)
   end
-  
+
   def confirmed?
     self.password_hash != nil
   end
-  
+
   def generate_reset_key
     reset_key = SecureRandom.urlsafe_base64
     #To consider: include salt?
@@ -47,7 +52,7 @@ class WebUser < ActiveRecord::Base
   end
 
   def self.authenticate(email, password)
-    web_user = where(:email => email).first
+    web_user = where(email: email).first
     if web_user && web_user.password_hash == web_user.hash_password(password)
       web_user
     else
