@@ -13,13 +13,13 @@
 
 ActiveRecord::Schema.define(version: 20130925001233) do
 
-  create_table "groups", force: true do |t|
+  create_table "groups", force: :cascade do |t|
     t.string "name", null: false
   end
 
   add_index "groups", ["name"], name: "index_groups_on_name", unique: true
 
-  create_table "job_summaries", force: true do |t|
+  create_table "job_summaries", force: :cascade do |t|
     t.integer "user_id",      null: false
     t.integer "system_id",    null: false
     t.date    "date",         null: false
@@ -32,7 +32,7 @@ ActiveRecord::Schema.define(version: 20130925001233) do
   add_index "job_summaries", ["date", "user_id", "system_id", "command_name"], name: "identity", unique: true
   add_index "job_summaries", ["date"], name: "index_job_summaries_on_date"
 
-  create_table "jobs", force: true do |t|
+  create_table "jobs", force: :cascade do |t|
     t.integer  "user_id",                 null: false
     t.integer  "system_id",               null: false
     t.string   "command_name", limit: 24, null: false
@@ -45,26 +45,25 @@ ActiveRecord::Schema.define(version: 20130925001233) do
   end
 
   add_index "jobs", ["command_name"], name: "index_jobs_on_command_name"
-  add_index "jobs", ["end_time"], name: "index_jobs_on_end_time"
   add_index "jobs", ["exit_code"], name: "index_jobs_on_exit_code"
-  add_index "jobs", ["start_time"], name: "index_jobs_on_start_time"
+  add_index "jobs", ["start_time", "end_time"], name: "index_jobs_on_start_time_and_end_time"
   add_index "jobs", ["system_id"], name: "index_jobs_on_system_id"
   add_index "jobs", ["user_id"], name: "index_jobs_on_user_id"
 
-  create_table "locks", force: true do |t|
-    t.string "name"
+  create_table "locks", force: :cascade do |t|
+    t.string "name", null: false
   end
 
   add_index "locks", ["name"], name: "index_locks_on_name", unique: true
 
-  create_table "system_types", force: true do |t|
+  create_table "system_types", force: :cascade do |t|
     t.string  "name",                       null: false
     t.integer "memory_stat_type", limit: 1, null: false
   end
 
   add_index "system_types", ["name"], name: "index_system_types_on_name", unique: true
 
-  create_table "systems", force: true do |t|
+  create_table "systems", force: :cascade do |t|
     t.string   "name",                     null: false
     t.integer  "system_type_id",           null: false
     t.datetime "start_time",               null: false
@@ -78,14 +77,14 @@ ActiveRecord::Schema.define(version: 20130925001233) do
   add_index "systems", ["start_time"], name: "index_systems_on_start_time"
   add_index "systems", ["system_type_id"], name: "index_systems_on_system_type_id"
 
-  create_table "users", force: true do |t|
+  create_table "users", force: :cascade do |t|
     t.string  "name",     null: false
     t.integer "group_id", null: false
   end
 
   add_index "users", ["name", "group_id"], name: "index_users_on_name_and_group_id", unique: true
 
-  create_table "web_users", force: true do |t|
+  create_table "web_users", force: :cascade do |t|
     t.string   "email",          null: false
     t.string   "password_hash"
     t.string   "password_salt"
